+++ RELEASE_NOTES.rst
@@ -9,6 +9,7 @@ ARTIQ-9 (Unreleased)
 * Hardware support:
    - 12Gbps CoaXPress grabber support on Kasli-SoC with CoaXPress-SFP adapter
      (and ZC706 with Hello-FPGA CXP 4R FMC card).
+   - Songbird - a 4-tone, 16-bit DDS using a 2.5Gsps LTC2000 DAC.
    - Improved SDRAM memory controller and DMA cores puts Kasli DMA performance on par with
      other platforms.
    - Core device reflashing over the network through the new ``flash`` tool in ``artiq_coremgmt``.